+++ app/helpers/spreadsheets_helper.rb
@@ -20,14 +20,23 @@
 
 module SpreadsheetsHelper
   def table_select_options
-    return project_tables.select(:id, :name) unless project_type_id
+    return project_tables.select(:id, :name) unless project_type_id || project_type_master
+    return project_type_tables.select(:id, :name) if project_type_id
 
-    project_type_tables.select(:id, :name)
+    project_type_master_tables.select(:id, :name)
   end
 
-  def project_type_tables
-    project_type.tables
+  def project_type_master
+    return unless defined? ProjectType
+
+    @project.is_project_type
   end
+
+  def project_type_master_tables
+    @project.tables
+  end
+
+  delegate :tables, to: :project_type, prefix: true
 
   def project_tables
     Table