+++ app/models/spreadsheet.rb
@@ -21,14 +21,16 @@
 class Spreadsheet < ActiveRecord::Base
   include Redmine::SafeAttributes
   include TableCalculation::Copyable
+  include TableCalculation::Sortable
 
-  belongs_to :project, foreign_key: :project_id, inverse_of: :spreadsheets
+  belongs_to :project, inverse_of: :spreadsheets
   belongs_to :table, inverse_of: :spreadsheets
   belongs_to :author, class_name: 'User'
   has_many :rows, class_name: 'SpreadsheetRow', dependent: :destroy
 
-  validates_uniqueness_of :name, scope: :project_id
-  validates_presence_of :name, :table_id
+  validates :name, uniqueness: { scope: :project_id }
+  validates :name, presence: true
+  validates :table_id, presence: true # table is not required when not validated!
 
   safe_attributes(
     :name,
@@ -38,9 +40,7 @@ class Spreadsheet < ActiveRecord::Base
     :author_id
   )
 
-  def column_ids
-    secure_table.column_ids
-  end
+  delegate :column_ids, to: :secure_table
 
   def calculations?
     secure_table.calculations.present?
@@ -61,7 +61,7 @@ class Spreadsheet < ActiveRecord::Base
   end
 
   def copy_rows(attributes = {})
-    rows.map do |row|
+    sorted_rows.map do |row|
       row.copy(attributes)
     end
   end
@@ -76,12 +76,22 @@ class Spreadsheet < ActiveRecord::Base
     all_row_values = spreadsheet.copied_row_values
     return unless all_row_values.any?(&:present?)
 
-    rows.zip(all_row_values).each do |row, row_values|
+    sorted_rows.zip(all_row_values).each do |row, row_values|
       row.assign_values(row_values)
     end
   end
 
   def copied_row_values
-    rows.map(&:copy_values)
+    sorted_rows.map(&:copy_values)
+  end
+
+  ##
+  # Rows needs to be sorted this way to avoid to copy them
+  # unsorted.
+  # @note There was an attempt to sort the rows in the association above
+  #       but that leads to side effects and ends in an exception.
+  #
+  def sorted_rows
+    sorted_by_id(rows)
   end
 end