+++ app/models/table.rb
@@ -36,8 +36,8 @@ class Table < ActiveRecord::Base
 
   has_many :calculations, dependent: :destroy, inverse_of: :table
 
-  validates_presence_of :name
-  validates_uniqueness_of :name
+  validates :name, presence: true
+  validates :name, uniqueness: true
 
   safe_attributes(
     :name,