+++ db/migrate/04_create_calculation_results.rb
@@ -20,19 +20,19 @@
 
 class CreateCalculationResults < ActiveRecord::Migration[4.2]
   def self.up
-    unless table_exists?(:calculation_results)
-      create_table :calculation_results do |t|
-        t.string :type, limit: 30, default: '', null: false
-        t.string :name, limit: 60, default: '', null: false
-        t.integer :calculation_id, default: 0, null: false
-        t.text :comment
-        t.timestamp :created_on
-        t.timestamp :updated_on
-      end
+    return if table_exists?(:calculation_results)
 
-      add_index :calculation_results, %i[calculation_id], name: 'calculation_results_by_calculation'
-      add_index :calculation_results, %i[type calculation_id], name: 'calculation_results_by_calculation_and_type'
+    create_table :calculation_results do |t|
+      t.string :type, limit: 30, default: '', null: false
+      t.string :name, limit: 60, default: '', null: false
+      t.integer :calculation_id, default: 0, null: false
+      t.text :comment
+      t.timestamp :created_on
+      t.timestamp :updated_on
     end
+
+    add_index :calculation_results, %i[calculation_id], name: 'calculation_results_by_calculation'
+    add_index :calculation_results, %i[type calculation_id], name: 'calculation_results_by_calculation_and_type'
   end
 
   def self.down