+++ lib/table_calculation/sortable.rb
@@ -2,9 +2,9 @@
 
 # This file is part of the Plugin Redmine Table Calculation.
 #
-# Copyright (C) 2021 - 2022 Liane Hampe <liaham@xmera.de>, xmera.
+# Copyright (C) 2022 Liane Hampe <liaham@xmera.de>, xmera.
 #
-# This program is free software; you can redistribute it and/or
+# This plugin program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
 # as published by the Free Software Foundation; either version 2
 # of the License, or (at your option) any later version.
@@ -19,17 +19,14 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
 module TableCalculation
-  ##
-  # Provide user login test
-  #
-  module CreateProjectType
-    def project_type_create_params(associates)
-      { project_type:
-        { name: 'Lore ipsum',
-          description: 'for testing',
-          is_public: 0,
-          default_member_role_id: 3,
-          position: 4 }.merge(associates) }
+  module Sortable
+    ##
+    # Will sort enumeration objects by its id.
+    # @examples
+    #  - ActiveRecord::Associations::CollectionProxy
+    #
+    def sorted_by_id(enumerations)
+      enumerations.sort_by(&:id)
     end
   end
 end