+++ test/project_type_creator.rb
@@ -44,5 +44,14 @@ module TableCalculation
         identifier: name.parameterize,
         is_project_type: true }
     end
+
+    def project_type_create_params(associates)
+      { project_type:
+        { name: 'Lore ipsum',
+          description: 'for testing',
+          is_public: 0,
+          default_member_role_id: 3,
+          position: 4 }.merge(associates) }
+    end
   end
 end