+++ test/test_helper.rb
@@ -28,7 +28,9 @@ require_relative 'load_fixtures'
 require_relative 'authenticate_user'
 require_relative 'enumerations'
 require_relative 'project_type_creator'
+require_relative 'project_creator'
 require_relative 'test_object_creators'
+require_relative 'prepare_spreadsheet'
 
 # The gem minitest-reporters gives color to the command-line
 require 'minitest/reporters'