+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.5 - 2022-12-09
+
+### Fixed
+
+* data set order changes
+* exceptions when copying table with certain settings
+* table selection field to show only relevant tables
+
 ## 1.0.4 - 2022-10-19
 
 ### Fixed